requirements.txt
coverage==4.5.1
Django==2.2a1
django-appconf==1.0.2
django-debug-toolbar==1.10.1
django-imagekit==4.0.2
django-markdownx==2.0.28
django-model-utils==3.1.2
django-notifications-hq==1.5.0
easy-thumbnails==2.6
gunicorn==19.9.0
jsonfield==2.0.2
Markdown==3.0.1
pilkit==2.0
Pillow==5.3.0
pytz==2018.7
six==1.11.0
sqlparse==0.2.4
whitenoise==4.1
